+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## main
+
+ - Add unprocessed waveform to summary plots.
  - Add config option to enable/disable STREC.
  - Add method to retrieve rupture model geometry info from workspace.
  - Add clipping probability as a trace processing parameter; remove it from failure message.